# [aws] cache user identity by 'aws configure list'