<!-- Describe the changes in this PR -->

Cache `aws sts get-caller-identity` result, indexed by hashing the output of `aws configure list`. This will speed up `sky launch` by ~1s in subsequent runs.

original discussion: https://github.com/skypilot-org/skypilot/issues/3159#issuecomment-2558120746

A simple benchmark:

```bash
# master, avg: 2.558s 
$ /usr/bin/time bash -c 'for i in {1..10}; do SKYPILOT_DISABLE_USAGE_COLLECTION=1 python -m sky.cli launch --gpus H100:8> /dev/null; done;'
       25,58 real        16,57 user        25,29 sys

# PR branch, avg: 1.596s
$ /usr/bin/time bash -c 'for i in {1..10}; do SKYPILOT_DISABLE_USAGE_COLLECTION=1 python -m sky.cli launch --gpus H100:8> /dev/null; done;'
       15,96 real        16,84 user        24,06 sys

# PR branch with https://github.com/skypilot-org/skypilot/pull/4483 picked, avg: 1.321s
$ /usr/bin/time bash -c 'for i in {1..10}; do SKYPILOT_DISABLE_USAGE_COLLECTION=1 python -m sky.cli launch --gpus H100:8> /dev/null; done;'
       13,21 real        16,28 user        11,18 sys
```

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: `bash format.sh`
- [x] Any manual or new tests for this PR (see above benchmark)
- [ ] All smoke tests: `pytest tests/test_smoke.py` 
- [ ] Relevant individual smoke tests: `pytest tests/test_smoke.py::test_fill_in_the_name` 
- [ ] Backward compatibility tests: `conda deactivate; bash -i tests/backward_compatibility_tests.sh`