@Michaelvll 2024-12-30
I assume the time tests listed in the PR description has the `launch` to return immediately after the optimizer has been run, i.e. not going to the confirmation line.