@aylei 2025-01-02
> I assume the time tests listed in the PR description has the `launch` to return immediately after the optimizer has been run, i.e. not going to the confirmation line.

Yes, I updated the PR description to make the result clearer~